Fixes ordering on sales
